// ideqs/src/models/userDataModel.ts
export interface UserDataModel {
  id?: number
  idAws?: string
  rol?: number
  user?: string
  tipoIdentificacion?: string
  identificacion?: string
  idSucursal?: number
  name?: string
  photo?: string
  sucursal?: string
  empresa?: string
  cargo?: string
  idEmpresa?: number
  born?: string
  city?: string
  correo?: string
  direction?: string
  lastname?: string
  activo?: number
  idCliente?: number
  logo?: string
  nombreLogo?: string
  direccion?: string
  telefono?: string
  password?: string
  picture?: File | null
  nombre?: string
  nit?: number
  pdf?: number
  physical?: number
  mobile?: number
  deleted?: number
  idUser?: number
}